Add functionality for syncing player lists
Added functionality for all the players that join a room to be able to see all of the other players that are part of the room.
This also ties in with making sure that players that join a room are indeed in the same multiplayer game room on the server.